LayoutTests/fast/events/spatial-navigation/snav-simple-content-overflow-expected.txt

It has a visible link_1.










... and an overflowed link like this.








This link should NOT get focused.

PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
